backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
distro==1.9.0
Django==5.1.2
django-environ==0.11.2
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
Markdown==3.7
openai==1.51.2
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-json-logger==2.0.7
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.6
tenacity==9.0.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.1
